Moved ExtensionEnvAdaptor into separate file
